Fixes general config issues with yelp api Does not fix yelp api call yet
